requirements.txt
asgiref==3.4.1
asn1crypto==0.24.0
attrs==21.2.0
autobahn==21.2.1
boto3==1.18.14
botocore==1.21.14
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2
defusedxml==0.7.1
dj-rest-auth==2.1.9
Django==3.2.5
django-allauth==0.45.0
django-appconf==1.0.5
django-cors-headers==3.8.0
django-debug-toolbar==3.2.1
django-imagekit==4.0.2
django-model-utils==4.1.1
django-request-logging==0.7.3
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
environs==9.3.3
GDAL==2.2.2
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==2.0.1
marshmallow==3.13.0
numpy==1.13.3
oauthlib==3.1.1
packaging==21.0
pilkit==2.0
Pillow==8.3.1
psycopg2-binary==2.9.1
pycparser==2.20
pycrypto==2.6.1
PyGObject==3.26.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.25
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
SecretStorage==2.3.1
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
uWSGI==2.0.19.1
zipp==3.5.0
zope.event==4.5.0
zope.interface==5.4.0
